requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
peewee==3.16.0
platformdirs==3.1.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
schedule==1.1.0
urllib3==1.26.15
